+++ domains/eventEditor/src/ui/EventRegistrationOptions/MaxRegistrations.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { __ } from '@eventespresso/i18n';
-import { GridItem, InlineEditText } from '../../';
+import { GridItem, InlineEditText } from '@eventespresso/components';
 import type { RegistrationOptionsMetaBoxProps } from './types';
 
 interface Props extends Pick<RegistrationOptionsMetaBoxProps, 'maxReg' | 'onMaxRegChange'> {}
