+++ domains/eventEditor/src/ui/EventRegistrationOptions/style.scss
