--- packages/components/src/EDTR/index.ts
@@ -1,1 +1,0 @@
-export * from './RegistrationOptionsMetaBox';
